:recycle: Move box to toolbox
