+++ src/components/toolbox/box/box.css
@@ -1,5 +1,5 @@
-@import '../app/variables.css';
-@import '../app/mixins.css';
+@import '../../app/variables.css';
+@import '../../app/mixins.css';
 
 :root {
   --header-height: 48px;
