+++ src/components/toolbox/box/content.js
