+++ src/components/toolbox/box/footerButton.css
@@ -1,5 +1,5 @@
-@import '../app/variables.css';
-@import '../app/mixins.css';
+@import '../../app/variables.css';
+@import '../../app/mixins.css';
 
 .footerButton {
   @mixin contentNormal bold;
